refactor: Use kebab-case for filenames
