+++ app/components/debug-panel.tsx
